RED - Add test for vending machine's initial state
This starts the outside-in TDD process by identifying the concepts in the first
feature regarding the vending machine's initial state. Ideally, the method
should be called #display, but I don't want to overwrite Object#display, an
inbuilt Ruby method. Already at this point I'm thinking about the fact that the
vending machine needs to be initialised in this state, and the fact that if I'm
not careful, VendingMachine#message will become a getter method.