RED - Add test for inserting nickel
Using language that's as close to real life, I write a test that tries to
model this sentence in the feature: "When a valid coin is inserted the amount
of the coin will be added to the current amount and the display will be
updated". Note two things: 1. The #message method returns a string, even when
displaying a monetary amount. 2. At this point I don't care about the
implementation of the coin, and use a double in its place.