GREEN - Accept a dime
Back to green. As suggested in the spec, coins do not know their value, and the
machine is responsible for assigning them a value. So, the coins have been
given a weight, based on their real-life weights, by having the doubles
respond to "weight".
Although this code passes the test, there are several things that I'm not happy
with - firstly, we have a whole bunch of magic numbers that don't explain
anything about the domain we're dealing with (in this case, coins). Also, to
someone who hasn't had their first cup of coffee, the %.2f conversion of floats
to strings with a certain degree of rounding is anything but descriptive.
However, I'm going to hold off refactoring until all types of coins are
handled, in case I learn something new.