REFACTOR - Move magic numbers to constants
weights_to_values has now been moved out into two hashes, which means that
what's going on is clearer. We also have finally identified the concept of
nickels, dimes and quarters in the code. I'm still not happy with the insert
method, as it violates SRP pretty heavily - it selects the correct coin AND
formats it, giving it two reasons to change. However, if I try to perform
Extract Method, I have to pass price around as an argument, which suggests that
there's something wrong with the design. For now, I'll leave it, because it
feels like I don't have enough information.