GREEN - Reject all other coins
Tests are green, but now I definitely want to refactor - value is definitely
going to get hit with Replace Temp with Query, and I somehow want to avoid the
explicit nil check. I'm also starting to suspect that @display is not the right
name for the state I'm holding, but I'm going to hold off that until a test
forces me to change that.