GREEN - Add coins together
Adding this test has introduced the idea of credit, which keeps track of the
value of the coins that have been inserted. It also means that I don't have to
pass around the coin as an argument quite as much, which makes me glad that
I didn't prematurely extract an object based on that, since it looks like
display and value extraction are based on different things (I already kinda
suspected that, but this has confirmed it).