REFACTOR - Remove display instance variable
I can eliminate the display instance variable, as it becomes clear that the
display depends on the credit in the machine. Also, I realised that I forgot to
remove the context block in my tests that read "when inserting valid coins", as
really all of these tests are to do with inserting coins in general.