GREEN - Place invalid coins in the coin return
Tests green, but there are a bunch of things I don't like here. I followed the
error messages one by one which is why I have a #coin_return method. However,
seeing as this is just a getter for @rejected_coins, in the refactor step I'll
probably collapse that into an attr_reader. I also still don't like the check
for nil. So let's see if I can fix that in the refactor.