REFACTOR - Eliminate nilcheck: insert coin complete
A few things renamed and moved around. By using fetch, we can provide a default
value of 0 for any coins that don't have an entry in COIN_WEIGHTS, eliminating
the need for a conditional. This may well come back to haunt me in the future
but for now it seems like the right thing to do. Also, the magic string "INSERT
COIN HERE" is extracted, and credit is now accessed through a private
attr_reader